Update unit test to reflect refactor
