+++ lib/ramble/ramble/spack_runner.py
@@ -81,7 +81,7 @@ class SpackRunner(object):
         self.shell = shell
 
         if self.shell == 'bash':
-            script = 'setup-env.py'
+            script = 'setup-env.sh'
         elif self.shell == 'csh':
             script = 'setup-env.csh'
         elif self.shell == 'fish':
