+++ lib/ramble/ramble/test/application_tests.py
@@ -72,106 +72,6 @@ def test_basic_app(mutable_mock_apps_repo):
         == 'Example var'
 
 
-def test_env_var_set_command_gen(mutable_mock_apps_repo):
-    basic_inst = mutable_mock_apps_repo.get('basic')
-
-    tests = {
-        'var1': 'val1',
-        'var2': 'val2'
-    }
-
-    answer = [
-        'export var1=val1;',
-        'export var2=val2;'
-    ]
-
-    out_cmds, _ = basic_inst._get_env_set_commands(tests, set())
-    for cmd in answer:
-        assert cmd in out_cmds
-
-
-def test_env_var_append_command_gen(mutable_mock_apps_repo):
-    basic_inst = mutable_mock_apps_repo.get('basic')
-
-    tests = [
-        {
-            'var-separator': ',',
-            'vars': {
-                'var1': 'val1',
-                'var2': 'val2'
-            },
-            'paths': {
-                'path1': 'path1',
-                'path2': 'path2'
-            }
-        },
-        {
-            'var-separator': ',',
-            'vars': {
-                'var1': 'val2',
-                'var2': 'val1'
-            },
-        }
-    ]
-
-    answer = [
-        'export var1="${var1},val1,val2";',
-        'export var2="${var2},val2,val1";',
-        'export path1="${path1}:path1";',
-        'export path2="${path2}:path2";'
-    ]
-
-    out_cmds, _ = basic_inst._get_env_append_commands(tests, set())
-    for cmd in answer:
-        assert cmd in out_cmds
-
-
-def test_env_var_prepend_command_gen(mutable_mock_apps_repo):
-    basic_inst = mutable_mock_apps_repo.get('basic')
-
-    tests = [
-        {
-            'paths': {
-                'path1': 'path1',
-                'path2': 'path2'
-            }
-        },
-        {
-            'paths': {
-                'path1': 'path2',
-                'path2': 'path1'
-            }
-        }
-    ]
-
-    answer = [
-        'export path1="path2:path1:${path1}";',
-        'export path2="path1:path2:${path2}";'
-    ]
-
-    out_cmds, _ = basic_inst._get_env_prepend_commands(tests, set())
-    for cmd in answer:
-        assert cmd in out_cmds
-
-
-def test_env_var_unset_command_gen(mutable_mock_apps_repo):
-    basic_inst = mutable_mock_apps_repo.get('basic')
-
-    tests = [
-        'var1',
-        'var2'
-    ]
-
-    answer = [
-        'unset var1;',
-        'unset var2;'
-    ]
-
-    out_cmds, _ = basic_inst._get_env_unset_commands(tests, set())
-    for cmd in answer:
-        assert cmd in out_cmds
-
-
 @pytest.mark.parametrize('app_name', ['basic', 'zlib'])
 def test_application_copy_is_deep(mutable_mock_apps_repo, app_name):
     src_inst = mutable_mock_apps_repo.get(app_name)
