+++ lib/ramble/ramble/test/end_to_end/explicit_zips.py
@@ -182,8 +182,6 @@ licenses:
 
             with open(os.path.join(exp_dir, 'full_command'), 'r') as f:
                 data = f.read()
-                # Test the license exists
-                assert "export WRF_LICENSE=port@server" in data
 
                 # Test the required environment variables exist
                 assert 'export OMP_NUM_THREADS="1"' in data
@@ -201,8 +199,6 @@ licenses:
 
             with open(os.path.join(exp_dir, 'execute_experiment'), 'r') as f:
                 data = f.read()
-                # Test the license exists
-                assert "export WRF_LICENSE=port@server" in data
 
                 # Test the required environment variables exist
                 assert 'export OMP_NUM_THREADS="1"' in data
