+++ lib/ramble/ramble/test/end_to_end/wrfv4_dry_run.py
@@ -175,10 +175,11 @@ licenses:
             assert os.path.exists(os.path.join(exp_dir, 'execute_experiment'))
             assert os.path.exists(os.path.join(exp_dir, 'full_command'))
 
+            license_inc_path = os.path.join(ws1.root, 'shared', 'licenses', 'wrfv4', 'license.inc')
             with open(os.path.join(exp_dir, 'full_command'), 'r') as f:
                 data = f.read()
                 # Test the license exists
-                assert "export WRF_LICENSE=port@server" in data
+                assert f". {license_inc_path}" in data
 
                 # Test the required environment variables exist
                 assert 'export OMP_NUM_THREADS="1"' in data
@@ -194,10 +195,13 @@ licenses:
                 # Test the run script has a reference to the experiment log file
                 assert os.path.join(exp_dir, f'{exp}.out') in data
 
+            with open(license_inc_path) as f:
+                data = f.read()
+                # Test the license is added to the include file
+                assert "export WRF_LICENSE=port@server" in data
+
             with open(os.path.join(exp_dir, 'execute_experiment'), 'r') as f:
                 data = f.read()
-                # Test the license exists
-                assert "export WRF_LICENSE=port@server" in data
 
                 # Test the required environment variables exist
                 assert 'export OMP_NUM_THREADS="1"' in data
